Remove duplicated detection logic from GetProp modules
