+++ mvt/android/modules/bugreport/getprop.py
@@ -4,15 +4,15 @@
 #   https://license.mvt.re/1.1/
 
 import logging
-from datetime import datetime, timedelta
 from typing import Optional
 
+from mvt.android.modules.detection_mixins import GetPropDetectionMixin
 from mvt.android.parsers import parse_getprop
 
 from .base import BugReportModule
 
 
-class Getprop(BugReportModule):
+class Getprop(GetPropDetectionMixin, BugReportModule):
     """This module extracts device properties from getprop command."""
 
     def __init__(
@@ -61,17 +61,4 @@ class Getprop(BugReportModule):
             lines.append(line)
 
         self.results = parse_getprop("\n".join(lines))
-
-        # Alert if phone is outdated.
-        for entry in self.results:
-            if entry["name"] == "ro.build.version.security_patch":
-                security_patch = entry["value"]
-                patch_date = datetime.strptime(security_patch, "%Y-%m-%d")
-                if (datetime.now() - patch_date) > timedelta(days=6 * 30):
-                    self.log.warning(
-                        "This phone has not received security updates "
-                        "for more than six months (last update: %s)",
-                        security_patch,
-                    )
-
         self.log.info("Extracted %d Android system properties", len(self.results))
